Reference

Infernal.cmd — Method

cmd(name, [args...])

Returns a command to call an Infernal binary name.

Infernal.exe — Method

exe(name, [args...])

Runs the Infernal binary name with arguments args.

Infernal.path — Method

path(name)

Path to Infernal binary name.

Infernal.Protocols.align_to_cm — Function

align_to_cm(fasta, cm, dir = dirname(fasta))

Aligns input sequences fasta to a covariance model cm, discarding bad matches. Follows Alternate Protocol from doi:10.1002/cpbi.51. Uses dir as working directory. Output files are prefixed with the name of the fasta file and the family name from cm.

Infernal.Protocols.emit_and_score_sequences — Function

emit_and_score_sequences(cm, dir = dirname(cm); N = 1, β = 1)

Emit N Infernal generated sequences based on the covariance model cm, at "inverse temperature" β. Uses dir as working directory.

Infernal.Protocols.extract_family_cm — Function

extract_family_cm(family, Rfamcm, dir = ".")

Extract the covariance model of family from the Rfam.cm (Rfamcm) file and write it to dir/family.cm.